+++ examples/util.js
@@ -168,5 +168,3 @@ export function getRadioSelectKeys(data, selectedKeys, key) {
   });
   return res;
 }
-
-// export { gData, getFilterExpandedKeys, getRadioSelectKeys };